refactor: PlanetTagEditor should be in components
